Pass in WriteBatch by reference and add `clear` method.
If you use WriteBatch'es within your project as kind of a buffer you
occasionally flush into RocksDB, it might be part of a struct. It's
difficult to move a single field out of a struct. Hence it is easier
if the methods on WriteBatch would take a WriteBatch by reference
and you call `clear()` on it manually once done.