Expose the `status` method on iterators
This allows handling errors when they're encountered during iteration of
the DB. The details of RocksDB iterator error reporting is described in
https://github.com/facebook/rocksdb/wiki/Iterator#error-handling

This method alone is not especially ergonomic or idiomatic for rust:
users would have to call `status` on their iterators whenever they
receive a `None` item to check whether iteration actually failed instead
of simply ended. A more complete approach is to change the iterator
signature to `Item = Result<..., Error>` and internally have the
iterator check its status. This better reflects the fact that an error
can be encountered during iteration, and that the user should handle or
propagate them explicitly. Such an approach would however be a breaking
change, and should be approached as a larger undertaking than this
changeset's simple addition.