Add get_updates_since() wrapper and its iterator
In concert with the recently added latest_sequence_number(), this returns an
iterator for the write batches that have occurred since the given sequence.
This API from RocksDB is based on callbacks, and as a result the wrapper
relies on the application implementing a trait that acts as a callback.

cargo test passes