Make set_iterate_upper_bound safe
This function is probably the only unsafe public API of this library.
The unsafe version is not super easy to use since it requires the user
to carefully make sure the buffer is alive. For example, the following
usage would be incorrect:

```
let mut opts = ReadOptions::default();
let upper_bound = b"k4".to_vec();
unsafe {
    opts.set_iterate_upper_bound(upper_bound);
}
```

A relatively easy way to make it safe is to store the buffer inside
`ReadOptions` and store `ReadOptions` inside the iterator, so the buffer
will not be dropped when the iterator is being used.

Fixes https://github.com/rust-rocksdb/rust-rocksdb/issues/299
Fixes https://github.com/rust-rocksdb/rust-rocksdb/issues/329