Fix lifetime parameter for iterators
The current implementation says that the lifetime of the iterator
(parameter 'b) lives at least as long as the lifetime of the DB
(parameter 'a), which is the opposite of what we want. Make it 'a: 'b
instead.

Also add lifetime parameters to a few extra iterator related functions.

Fixes https://github.com/rust-rocksdb/rust-rocksdb/issues/386