Removed unneeded bindgen features
Some features of bindgen are not needed when building the library and
lead to unused dependencies (mainly `clap`). This change deactivates all
`bindgen` features except for `runtime` which is required for build to
work.

This improved debug build time by ~13 seconds and (surprisingly) release
build time by ~9 seconds. But perhaps improved security is more
interesting.