Allow passing &CStr arguments
Methods such as property_value take `&str` name as argument.  However,
since they interact with C functions, such string arguments need to be
converted into a `&CStr`.  Because the argument is `&str` this must be
done via `CString::new` which needs to allocate memory and verify that
the value does not have embedded NUL bytes.  Even if the caller
already has a `&CStr` there is no way to pass it directly to the
function and a needless allocation happens.

This commit introduces a `CStrLike` trait which can be used in place
of `&str` arguments.  A function which accepts `impl CStrLike` as an
argument can be called either with `&str` or `&CStr` which allows it
to take the C string directly if it’s passed.

Note that this is strictly speaking backwards-incompatible change.
While call sites such as `db.property_value("foo")` won’t be affected,
call sites which use long deref chains or need type inference will
stop compiling.

Alternatives to this are to:
1. do nothing,
2. introduce `_cstr` variants of functions which need `&CStr`
   internally or
3. change argument type to `&CStr` where functions do the conversion.

The first option forces users to incur the allocation and verification
cost each time method is called.  The second bloats the API by
introducing near-duplicates of existing methods and may lead to
further exponential explosion in the future.  The third one removes
convenience of using string literals from users who might not care
about performance as much.

I think the third option is actually the cleanest but wished to
maintain some backwards compatibility so went with this commit.