Use NonNull in DBRawIteratorWithThreadMode
The inner pointer in DBRawIteratorWithThreadMode is never null so mark
it as such.  With this change, Rust can optimise layout of the types
such that Option<DBRawIteratorWithThreadMode> is the same size as
DBRawIteratorWithThreadMode.  Previously it would be 8 bytes larger
(72 vs 64 bytes).

Ideally rocksdb_create_iterator would return NonNull but since it
doesn’t we need to go ahead and check for null explicitly.