Refactor iterator-related tests
Firstly, introduce helper assertions which reduce code repetition.
They do it by encapsulating verifying output of the iterator into
a single location.  This will be useful in upcoming commit which
changes the Item type of the iterator.

Secondly, keep keys and values declared at the start of some of the
tests as slices without boxing them.  Doing the latter introduces
conveluted syntax such as `&*k1`; keeping them as slices simplifies
the syntax and also allows them to be converte into constsants.

Lastly, replace `cba` helper methods with `pair` which construct
a pair of boxed slices.  This removes some repetition since after
the previous point `cba` was only ever used in a pair.