Wrap props names into a PropName type offering free conversion to str
Introduce properties::PropName and properties::PropertyName which wrap
CStr and CString types respectively with additional requirement that
the value is valid UTF-8.  This means that the types offer zero-cost
conversion into string.

Use it with property names in properties module such that the
properties can be used with property_value (which was the case before)
and also Displayed (which previously required UTF-8 verification or an
unsafe block and now can be done safely with zero additional cost).

This is API breaking change (since type of objects in properties
module change) but the breakage should be easy to deal with
(especially since the new types implement CStrLike and thus most
common usage of the property names will continue to work).