Expose rocksdb cumulative statistics and histograms
This exposes rocksdb statistics (tickers and histograms) to allow rust-rocksdb users to introspect individual metrics.
Additionally, this PR also adds `set_statistics_level(StatsLevel)` to `Options` for fine-grain control over which
stats are collected.

Tickers and histograms are available in `rocksdb::statistics::{Ticker, Histogram}` with the necessary helpers to
convert the ticker type to the corresponding rocksdb name (and vice-versa via the `FromStr` trait).

Test Plan:
- New test added in test_db.rs
- Sanity check unit tests in statistics.rs