fix: compaction filter non-static Change
Presently the compaction filter declare the Change case of the Decision enum as
holding 'static [u8] slice. That made it rather useless in safe Rust as any
non-trivial data change must leak the returned data.

To address that replace the static slice in the Change data with Vec<u8>. The
change callback implementation then stores this vector in the thread-local data
and return a pointer to it. This is safe as on C++ side RocksDB uses the
returned pointer just to copy the data pointed to it to C++ string immediately
after the callback call.

Then the code use various opportuneties to drop the the thread-local storage
like the following callback calls or the vallback destruction.

This is a breaking change as it changes the type of publically facing
Decision::Change payload. However upgrading the existing code should be trivial.
And in case the code does not use the Change no changes will be necessary.